Streamlit/YOLO_requirements.txt
# =========================
# Deep Learning Framework
# =========================
torch
torchvision

# =========================
# YOLOv8 (Ultralytics)
# =========================
ultralytics
opencv-python

# =========================
# Web App
# =========================
streamlit

# =========================
# Image & Numerical Processing
# =========================
Pillow
numpy
imagehash
scikit-image

# =========================
# Plotting / Visualization
# =========================
matplotlib
